Move extension to TestUtil namespace
